+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 # Changelog
 
+## [3.1.2] - 2026-01-14
+
+### Fixed
+
+- Changed error Markdown panel from `[!ERROR]` to `[!CAUTION]` as ERROR is not a
+  supported GitHub alert type
+- Fixed duplicate `@mention` escaping in PR comments if already escaped
+- Fixed file being relative to runner root instead of repository root in PR
+  comments
+- Fixed PR comments to be minimized when the latest run has no skipped
+  annotations, ensuring old comments are cleared when they're no longer relevant
+
+### Improved
+
+- Truncate long file paths in PR comments for better readability
+
 ## [3.1.1] - 2025-12-30
 
 ### Fixed